Clone meta-alb as meta-fsl-s32v directory
To ease up the transition from internal repositories to repositories hosted on
CAF, we first replace the meta-fsl-s32v repo with the meta-alb repo, but keep
the same name in the yocto source layout. meta-alb should be a drop-in
replacement anyway, so there should be no impact on the build.

Issue: ALB-1875, ALB-1876
Signed-off-by: Eddy Petrișor <eddy.petrisor@nxp.com>